move launch files into launch.legacy namespace
